Microbial Community Class

A class to represent communities of microorganisms.

Slots

species

A character vector representing the microorganisms present in the community.

metabolites

A character vector representing the metabolites consumed and produced within the community.

fluxes

A numeric vector representing the fluxes of each metabolite in the community.

edges

A list containing a hash of neighbouring metabolites and neighbouring species. For neighbouring metabolites, both the ingoing and outgoing flux is specified.

bin_matrix

A binary matrix representing the presence of fluxes.

name

A string representing the name of the community.

n_edges_matrix

A matrix representing the number of species in the edge between two metabolites.

flux_prod_j_matrix

A matrix representing the total flux of production of j in the community.

eff_flux_prod_j_matrix

A matrix giving the effective fluxes of production of j in the community.

flux_cons_i_matrix

A matrix giving the total flux of consumption of i in the community.

eff_flux_cons_i_matrix

A matrix giving the effective fluxes of consumption of i in the community.