Functional Microbiome Representation based on TreeSummarizedExperiment

Functional Microbiome Representation based on TreeSummarizedExperiment

Usage

MicrobiomeFunction(
  data,
  name = NA_character_,
  species_col = "species",
  metabolite_col = "met",
  flux_col = "flux",
  ...
)

Arguments

data

a DataFrame-like object that includes columns specfiying the species, metabolites and fluxes in the microbiome. The fluxes can either be weighted or unweighted (all of magnitude 1).

name

a character scalar specifying the name of the Microbiome

species_col

Character scalar specfiying the name of the species column, defaults to 'spec'.

metabolite_col

Character scalar specifying the name of the metabolite column, defaults to 'met'.

flux_col

Character scalar specifying the name of the flux column, defaults to 'flux'.

...

Additional arguments to be passed to the constructor.