Get All Edges of a Community

Takes a tibble describing a microbial community as input and returns a list with two hashs, one giving the edges between metabolites and the other between species.

Usage

findEdges(tb, silent = FALSE)

Arguments

tb

Tibble describing a microbial community

silent

Logical to toggle output

Value

A list of two hashs, one for edges between metabolites and one for species.

Examples

findEdges(ac_A1R12_1)
#> -------- A1R12 --------
#> Producers: ac ala__D ala__L co2 glu__L gthox 
#> Consumers: acald asp__L etoh gly gthrd h2o2 h2s pyr 
#> -------- I2R16 --------
#> Producers: acald asp__L co2 etoh gly gthrd h2o2 h2s pyr 
#> Consumers: ac ala__D ala__L glu__L gthox 
#> -------- ac --------
#> From: acald asp__L co2 etoh gly gthrd h2o2 h2s pyr 
#> To: acald asp__L etoh gly gthrd h2o2 h2s pyr 
#> -------- acald --------
#> From: ac ala__D ala__L co2 glu__L gthox 
#> To: ac ala__D ala__L glu__L gthox 
#> -------- ala__D --------
#> From: acald asp__L co2 etoh gly gthrd h2o2 h2s pyr 
#> To: acald asp__L etoh gly gthrd h2o2 h2s pyr 
#> -------- ala__L --------
#> From: acald asp__L co2 etoh gly gthrd h2o2 h2s pyr 
#> To: acald asp__L etoh gly gthrd h2o2 h2s pyr 
#> -------- asp__L --------
#> From: ac ala__D ala__L co2 glu__L gthox 
#> To: ac ala__D ala__L glu__L gthox 
#> -------- co2 --------
#> From:  
#> To: acald asp__L etoh gly gthrd h2o2 h2s pyr ac ala__D ala__L glu__L gthox 
#> -------- etoh --------
#> From: ac ala__D ala__L co2 glu__L gthox 
#> To: ac ala__D ala__L glu__L gthox 
#> -------- glu__L --------
#> From: acald asp__L co2 etoh gly gthrd h2o2 h2s pyr 
#> To: acald asp__L etoh gly gthrd h2o2 h2s pyr 
#> -------- gly --------
#> From: ac ala__D ala__L co2 glu__L gthox 
#> To: ac ala__D ala__L glu__L gthox 
#> -------- gthox --------
#> From: acald asp__L co2 etoh gly gthrd h2o2 h2s pyr 
#> To: acald asp__L etoh gly gthrd h2o2 h2s pyr 
#> -------- gthrd --------
#> From: ac ala__D ala__L co2 glu__L gthox 
#> To: ac ala__D ala__L glu__L gthox 
#> -------- h2o2 --------
#> From: ac ala__D ala__L co2 glu__L gthox 
#> To: ac ala__D ala__L glu__L gthox 
#> -------- h2s --------
#> From: ac ala__D ala__L co2 glu__L gthox 
#> To: ac ala__D ala__L glu__L gthox 
#> -------- pyr --------
#> From: ac ala__D ala__L co2 glu__L gthox 
#> To: ac ala__D ala__L glu__L gthox 
#> $metabolites
#> <hash> containing 2 key-value pair(s).
#>   A1R12 : <hash> containing 4 key-value pair(s).
#>       cons_fluxes : -1.12096390 -0.58268031 -0.18475279 -0.02158742 -2.90547696 -1.45273848 -0.00238294 -2.01918234
#>       consumers : acald  asp__L etoh   gly    gthrd  h2o2   h2s    pyr   
#>       prod_fluxes : 0.7729250 0.7603136 1.2168416 0.6184093 0.5068255 1.4527385
#>       producers : ac     ala__D ala__L co2    glu__L gthox 
#>   I2R16 : <hash> containing 4 key-value pair(s).
#>       cons_fluxes : -10.7673345  -0.7603136  -1.2168416  -0.5068255  -1.4527385
#>       consumers : ac     ala__D ala__L glu__L gthox 
#>       prod_fluxes :  1.12096390  0.58268031 12.12608512  0.18475279  0.02158742  2.90547696  1.45273848  0.00238294  2.01918234
#>       producers : acald  asp__L co2    etoh   gly    gthrd  h2o2   h2s    pyr   
#> 
#> $species
#> <hash> containing 14 key-value pair(s).
#>   ac : <hash> containing 2 key-value pair(s).
#>       from : acald  asp__L co2    etoh   gly    gthrd  h2o2   h2s    pyr   
#>       to : acald  asp__L etoh   gly    gthrd  h2o2   h2s    pyr   
#>   acald : <hash> containing 2 key-value pair(s).
#>       from : ac     ala__D ala__L co2    glu__L gthox 
#>       to : ac     ala__D ala__L glu__L gthox 
#>   ala__D : <hash> containing 2 key-value pair(s).
#>       from : acald  asp__L co2    etoh   gly    gthrd  h2o2   h2s    pyr   
#>       to : acald  asp__L etoh   gly    gthrd  h2o2   h2s    pyr   
#>   ala__L : <hash> containing 2 key-value pair(s).
#>       from : acald  asp__L co2    etoh   gly    gthrd  h2o2   h2s    pyr   
#>       to : acald  asp__L etoh   gly    gthrd  h2o2   h2s    pyr   
#>   asp__L : <hash> containing 2 key-value pair(s).
#>       from : ac     ala__D ala__L co2    glu__L gthox 
#>       to : ac     ala__D ala__L glu__L gthox 
#>   co2 : <hash> containing 2 key-value pair(s).
#>       from : 
#>       to : acald  asp__L etoh   gly    gthrd  h2o2   h2s    pyr    ac     ala__D ala__L glu__L gthox 
#>   etoh : <hash> containing 2 key-value pair(s).
#>       from : ac     ala__D ala__L co2    glu__L gthox 
#>       to : ac     ala__D ala__L glu__L gthox 
#>   glu__L : <hash> containing 2 key-value pair(s).
#>       from : acald  asp__L co2    etoh   gly    gthrd  h2o2   h2s    pyr   
#>       to : acald  asp__L etoh   gly    gthrd  h2o2   h2s    pyr   
#>   gly : <hash> containing 2 key-value pair(s).
#>       from : ac     ala__D ala__L co2    glu__L gthox 
#>       to : ac     ala__D ala__L glu__L gthox 
#>   gthox : <hash> containing 2 key-value pair(s).
#>       from : acald  asp__L co2    etoh   gly    gthrd  h2o2   h2s    pyr   
#>       to : acald  asp__L etoh   gly    gthrd  h2o2   h2s    pyr   
#>   gthrd : <hash> containing 2 key-value pair(s).
#>       from : ac     ala__D ala__L co2    glu__L gthox 
#>       to : ac     ala__D ala__L glu__L gthox 
#>   h2o2 : <hash> containing 2 key-value pair(s).
#>       from : ac     ala__D ala__L co2    glu__L gthox 
#>       to : ac     ala__D ala__L glu__L gthox 
#>   h2s : <hash> containing 2 key-value pair(s).
#>       from : ac     ala__D ala__L co2    glu__L gthox 
#>       to : ac     ala__D ala__L glu__L gthox 
#>   pyr : <hash> containing 2 key-value pair(s).
#>       from : ac     ala__D ala__L co2    glu__L gthox 
#>       to : ac     ala__D ala__L glu__L gthox 
#>