Get All Edges of a Community

Takes a tibble describing a microbial community as input and returns a list with two hashs, one giving the edges between metabolites and the other between species.

Usage

get_edges(tb, silent = FALSE)

Arguments

tb

Tibble describing a microbial community

silent

Logical to toggle output

Value

A list of two hashs, one for edges between metabolites and one for species.

Examples

get_edges(ac_A1R12_1)
#> -------- ac --------
#> Producers: B3M02 
#> Consumers: m_1A01 
#> -------- acald --------
#> Producers: B3M02 
#> Consumers: m_1A01 
#> -------- ade --------
#> Producers: B3M02 
#> Consumers: m_1A01 
#> -------- ala__D --------
#> Producers: m_1A01 
#> Consumers: B3M02 
#> -------- ala__L --------
#> Producers: m_1A01 
#> Consumers: B3M02 
#> -------- arg__L --------
#> Producers: m_1A01 
#> Consumers: B3M02 
#> -------- asp__L --------
#> Producers: m_1A01 
#> Consumers: B3M02 
#> -------- cgly --------
#> Producers: B3M02 
#> Consumers: m_1A01 
#> -------- co2 --------
#> Producers: m_1A01 
#> Consumers: B3M02 
#> -------- etoh --------
#> Producers: B3M02 
#> Consumers: m_1A01 
#> -------- f6p --------
#> Producers:  
#> Consumers: m_1A01 
#> -------- for --------
#> Producers: m_1A01 
#> Consumers: B3M02 
#> -------- glu__L --------
#> Producers: m_1A01 
#> Consumers: B3M02 
#> -------- gly --------
#> Producers: B3M02 
#> Consumers: m_1A01 
#> -------- glyald --------
#> Producers: m_1A01 
#> Consumers: B3M02 
#> -------- glyc --------
#> Producers: B3M02 
#> Consumers: m_1A01 
#> -------- gthrd --------
#> Producers: m_1A01 
#> Consumers: B3M02 
#> -------- h2s --------
#> Producers: B3M02 
#> Consumers: m_1A01 
#> -------- ile__L --------
#> Producers: B3M02 
#> Consumers: m_1A01 
#> -------- leu__L --------
#> Producers: m_1A01 
#> Consumers: B3M02 
#> -------- orn --------
#> Producers: B3M02 
#> Consumers: m_1A01 
#> -------- phe__L --------
#> Producers: m_1A01 
#> Consumers: B3M02 
#> -------- pro__L --------
#> Producers: B3M02 
#> Consumers: m_1A01 
#> -------- pyr --------
#> Producers: m_1A01 
#> Consumers: B3M02 
#> -------- ser__L --------
#> Producers: m_1A01 
#> Consumers: B3M02 
#> -------- thr__L --------
#> Producers: B3M02 
#> Consumers: m_1A01 
#> -------- trp__L --------
#> Producers: m_1A01 
#> Consumers: B3M02 
#> -------- tyr__L --------
#> Producers: m_1A01 
#> Consumers: B3M02 
#> -------- ura --------
#> Producers: m_1A01 
#> Consumers: B3M02 
#> -------- val__L --------
#> Producers: m_1A01 
#> Consumers: B3M02 
#> -------- B3M02 --------
#> From: m_1A01 m_1A01 m_1A01 m_1A01 m_1A01 m_1A01 m_1A01 m_1A01 m_1A01 m_1A01 m_1A01 m_1A01 m_1A01 m_1A01 m_1A01 m_1A01 m_1A01 
#> To: m_1A01 m_1A01 m_1A01 m_1A01 m_1A01 m_1A01 m_1A01 m_1A01 m_1A01 m_1A01 m_1A01 m_1A01 
#> -------- m_1A01 --------
#> From: B3M02 B3M02 B3M02 B3M02 B3M02 B3M02 B3M02 B3M02 B3M02 B3M02 B3M02 B3M02 
#> To: B3M02 B3M02 B3M02 B3M02 B3M02 B3M02 B3M02 B3M02 B3M02 B3M02 B3M02 B3M02 B3M02 B3M02 B3M02 B3M02 B3M02 
#> $metabolites
#> <hash> containing 30 key-value pair(s).
#>   ac : <hash> containing 4 key-value pair(s).
#>       cons_fluxes : -0.4910885
#>       consumers : m_1A01
#>       prod_fluxes : 0.4910885
#>       producers : B3M02
#>   acald : <hash> containing 4 key-value pair(s).
#>       cons_fluxes : -0.8734409
#>       consumers : m_1A01
#>       prod_fluxes : 0.8734409
#>       producers : B3M02
#>   ade : <hash> containing 4 key-value pair(s).
#>       cons_fluxes : -0.2735564
#>       consumers : m_1A01
#>       prod_fluxes : 0.2735564
#>       producers : B3M02
#>   ala__D : <hash> containing 4 key-value pair(s).
#>       cons_fluxes : -0.9687185
#>       consumers : B3M02
#>       prod_fluxes : 0.9687185
#>       producers : m_1A01
#>   ala__L : <hash> containing 4 key-value pair(s).
#>       cons_fluxes : -1.164551
#>       consumers : B3M02
#>       prod_fluxes : 1.164551
#>       producers : m_1A01
#>   arg__L : <hash> containing 4 key-value pair(s).
#>       cons_fluxes : -0.1455126
#>       consumers : B3M02
#>       prod_fluxes : 0.1455126
#>       producers : m_1A01
#>   asp__L : <hash> containing 4 key-value pair(s).
#>       cons_fluxes : -1.296748
#>       consumers : B3M02
#>       prod_fluxes : 1.296748
#>       producers : m_1A01
#>   cgly : <hash> containing 4 key-value pair(s).
#>       cons_fluxes : -0.4343578
#>       consumers : m_1A01
#>       prod_fluxes : 0.4343578
#>       producers : B3M02
#>   co2 : <hash> containing 4 key-value pair(s).
#>       cons_fluxes : -4.141632
#>       consumers : B3M02
#>       prod_fluxes : 4.145762
#>       producers : m_1A01
#>   etoh : <hash> containing 4 key-value pair(s).
#>       cons_fluxes : -3.077723
#>       consumers : m_1A01
#>       prod_fluxes : 3.077723
#>       producers : B3M02
#>   f6p : <hash> containing 4 key-value pair(s).
#>       cons_fluxes : -10
#>       consumers : m_1A01
#>       prod_fluxes : 
#>       producers : 
#>   for : <hash> containing 4 key-value pair(s).
#>       cons_fluxes : -0.9164858
#>       consumers : B3M02
#>       prod_fluxes : 0.9164858
#>       producers : m_1A01
#>   glu__L : <hash> containing 4 key-value pair(s).
#>       cons_fluxes : -0.3306133
#>       consumers : B3M02
#>       prod_fluxes : 0.3306133
#>       producers : m_1A01
#>   gly : <hash> containing 4 key-value pair(s).
#>       cons_fluxes : -4.405985
#>       consumers : m_1A01
#>       prod_fluxes : 4.405985
#>       producers : B3M02
#>   glyald : <hash> containing 4 key-value pair(s).
#>       cons_fluxes : -3.235951
#>       consumers : B3M02
#>       prod_fluxes : 3.235951
#>       producers : m_1A01
#>   glyc : <hash> containing 4 key-value pair(s).
#>       cons_fluxes : -1.728675
#>       consumers : m_1A01
#>       prod_fluxes : 1.728675
#>       producers : B3M02
#>   gthrd : <hash> containing 4 key-value pair(s).
#>       cons_fluxes : -0.4343578
#>       consumers : B3M02
#>       prod_fluxes : 0.4343578
#>       producers : m_1A01
#>   h2s : <hash> containing 4 key-value pair(s).
#>       cons_fluxes : -0.2250996
#>       consumers : m_1A01
#>       prod_fluxes : 0.2250996
#>       producers : B3M02
#>   ile__L : <hash> containing 4 key-value pair(s).
#>       cons_fluxes : -0.2657752
#>       consumers : m_1A01
#>       prod_fluxes : 0.2657752
#>       producers : B3M02
#>   leu__L : <hash> containing 4 key-value pair(s).
#>       cons_fluxes : -0.2216353
#>       consumers : B3M02
#>       prod_fluxes : 0.2216353
#>       producers : m_1A01
#>   orn : <hash> containing 4 key-value pair(s).
#>       cons_fluxes : -0.1455126
#>       consumers : m_1A01
#>       prod_fluxes : 0.1455126
#>       producers : B3M02
#>   phe__L : <hash> containing 4 key-value pair(s).
#>       cons_fluxes : -0.09113971
#>       consumers : B3M02
#>       prod_fluxes : 0.09113971
#>       producers : m_1A01
#>   pro__L : <hash> containing 4 key-value pair(s).
#>       cons_fluxes : -0.2022205
#>       consumers : m_1A01
#>       prod_fluxes : 0.2022205
#>       producers : B3M02
#>   pyr : <hash> containing 4 key-value pair(s).
#>       cons_fluxes : -3.97563
#>       consumers : B3M02
#>       prod_fluxes : 3.97563
#>       producers : m_1A01
#>   ser__L : <hash> containing 4 key-value pair(s).
#>       cons_fluxes : -0.5845611
#>       consumers : B3M02
#>       prod_fluxes : 0.5845611
#>       producers : m_1A01
#>   thr__L : <hash> containing 4 key-value pair(s).
#>       cons_fluxes : -0.2320722
#>       consumers : m_1A01
#>       prod_fluxes : 0.2320722
#>       producers : B3M02
#>   trp__L : <hash> containing 4 key-value pair(s).
#>       cons_fluxes : -0.02796348
#>       consumers : B3M02
#>       prod_fluxes : 0.02796348
#>       producers : m_1A01
#>   tyr__L : <hash> containing 4 key-value pair(s).
#>       cons_fluxes : -0.06783689
#>       consumers : B3M02
#>       prod_fluxes : 0.06783689
#>       producers : m_1A01
#>   ura : <hash> containing 4 key-value pair(s).
#>       cons_fluxes : -0.1627321
#>       consumers : B3M02
#>       prod_fluxes : 0.1627321
#>       producers : m_1A01
#>   val__L : <hash> containing 4 key-value pair(s).
#>       cons_fluxes : -0.2084547
#>       consumers : B3M02
#>       prod_fluxes : 0.2084547
#>       producers : m_1A01
#> 
#> $species
#> <hash> containing 2 key-value pair(s).
#>   B3M02 : <hash> containing 2 key-value pair(s).
#>       from : m_1A01 m_1A01 m_1A01 m_1A01 m_1A01 m_1A01 m_1A01 m_1A01 m_1A01 m_1A01 m_1A01 m_1A01 m_1A01 m_1A01 m_1A01 m_1A01 m_1A01
#>       to : m_1A01 m_1A01 m_1A01 m_1A01 m_1A01 m_1A01 m_1A01 m_1A01 m_1A01 m_1A01 m_1A01 m_1A01
#>   m_1A01 : <hash> containing 2 key-value pair(s).
#>       from : B3M02 B3M02 B3M02 B3M02 B3M02 B3M02 B3M02 B3M02 B3M02 B3M02 B3M02 B3M02
#>       to : B3M02 B3M02 B3M02 B3M02 B3M02 B3M02 B3M02 B3M02 B3M02 B3M02 B3M02 B3M02 B3M02 B3M02 B3M02 B3M02 B3M02
#>